pyproject.toml
[project]
name = "tg-bot-sol"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.12",
    "hpke>=0.3.2",
    "httpx>=0.28.1",
    "pyhpke>=0.6.2",
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.1",
    "requests>=2.32.4",
    "solana>=0.36.7",
]
